feat: get topic users from server instead of DOM
Mentions pre-populates the autocomplete dropdown with a list of
users in the topic. The original logic was rather naive and only
checked the DOM for users, but would not work if a topic spanned
multiple pages (or in the case of IS, if the posts had not been
loaded yet into the DOM).

This commit changes this behaviour so that the posters in the
topic are grabbed from the server instead.